Introduced RealmBase and SharedGroupManager
