+++ realm/src/androidTest/java/io/realm/RealmTest.java
@@ -202,15 +202,15 @@ public class RealmTest extends AndroidTestCase {
         Realm realm = null;
         try {
             realm = Realm.getInstance(realmConfig);
-            assertEquals(1, Realm.handlers.size());
+            assertEquals(1, Realm.getHandlers().size());
             realm.close();
 
             // All Realms closed. No handlers should be alive.
-            assertEquals(0, Realm.handlers.size());
+            assertEquals(0, Realm.getHandlers().size());
 
             // Open instance the 2nd time. Old handler should now be gone
             realm = Realm.getInstance(realmConfig);
-            assertEquals(1, Realm.handlers.size());
+            assertEquals(1, Realm.getHandlers().size());
             realm.close();
 
         } finally {
@@ -678,21 +678,6 @@ public class RealmTest extends AndroidTestCase {
             fail("Expected exception");
         } catch (IllegalStateException ignored) {
         }
-    }
-
-    // int getVersion() AND void setVersion(int version)
-    public void testGetVersionAndSetVersion() throws IOException {
-        // ** Initial version must be 0
-        populateTestRealm();
-        long version = testRealm.getVersion();
-        assertEquals(0, version);
-
-        // ** Version should be updateable
-        version = 42;
-        testRealm.beginTransaction();
-        testRealm.setVersion(version);
-        testRealm.commitTransaction();
-        assertEquals(version, testRealm.getVersion());
     }
 
     public void testShouldFailOutsideTransaction() {
@@ -1653,7 +1638,6 @@ public class RealmTest extends AndroidTestCase {
         try { testRealm.copyToRealmOrUpdate(ts);        fail(); } catch (IllegalStateException expected) {}
         try { testRealm.remove(AllTypes.class, 0);      fail(); } catch (IllegalStateException expected) {}
         try { testRealm.clear(AllTypes.class);          fail(); } catch (IllegalStateException expected) {}
-        try { testRealm.setVersion(42);                 fail(); } catch (IllegalStateException expected) {}
 
         try { testRealm.createObjectFromJson(AllTypesPrimaryKey.class, jsonObj);                fail(); } catch (RealmException expected) {}
         try { testRealm.createObjectFromJson(AllTypesPrimaryKey.class, jsonObjStr);             fail(); } catch (RealmException expected) {}
@@ -1673,38 +1657,47 @@ public class RealmTest extends AndroidTestCase {
     // TODO: re-introduce this test mocking the ReferenceQueue instead of relying on the GC
 /*    // Check that FinalizerRunnable can free native resources (phantom refs)
     public void testReferenceCleaning() throws NoSuchFieldException, IllegalAccessException {
-        Field sharedGroupReference = Realm.class.getDeclaredField("sharedGroup");
-        sharedGroupReference.setAccessible(true);
-        SharedGroup sharedGroup = (SharedGroup) sharedGroupReference.get(testRealm);
-        assertNotNull(sharedGroup);
 
+        RealmConfiguration config = new RealmConfiguration.Builder(getContext()).name("myown").build();
+        Realm.deleteRealm(config);
+        testRealm = Realm.getInstance(config);
+
+        // Manipulate field accessibility to facilitate testing
+        Field realmFileReference = RealmBase.class.getDeclaredField("sharedGroupManager");
+        realmFileReference.setAccessible(true);
         Field contextField = SharedGroup.class.getDeclaredField("context");
         contextField.setAccessible(true);
-        io.realm.internal.Context context = (io.realm.internal.Context) contextField.get(sharedGroup);
-        assertNotNull(context);
-
         Field rowReferencesField = io.realm.internal.Context.class.getDeclaredField("rowReferences");
         rowReferencesField.setAccessible(true);
+
+        SharedGroupManager realmFile = (SharedGroupManager) realmFileReference.get(testRealm);
+        assertNotNull(realmFile);
+
+        io.realm.internal.Context context = (io.realm.internal.Context) contextField.get(realmFile.getSharedGroup());
+        assertNotNull(context);
+
         List<Reference<?>> rowReferences = (List<Reference<?>>) rowReferencesField.get(context);
         assertNotNull(rowReferences);
-
 
         // insert some rows, then give the thread some time to cleanup
         // we have 8 reference so far let's add more
         final int numberOfPopulateTest = 1000;
-        final int totalNumberOfReferences = 8 + 20 * 2 * numberOfPopulateTest;
+        final int numberOfObjects = 20;
+        final int totalNumberOfReferences = 8 + numberOfObjects * 2 * numberOfPopulateTest;
 
         long tic = System.currentTimeMillis();
         for (int i = 0; i < numberOfPopulateTest; i++) {
-            populateTestRealm(testRealm, 20);
+            populateTestRealm(testRealm, numberOfObjects);
         }
         long toc = System.currentTimeMillis();
         Log.d(RealmTest.class.getName(), "Insertion time: " + (toc - tic));
 
         final int MAX_GC_RETRIES = 5;
         int numberOfRetries = 0;
+        Log.i("GCing", "Hoping for the best");
         while (rowReferences.size() > 0 && numberOfRetries < MAX_GC_RETRIES) {
             SystemClock.sleep(TimeUnit.SECONDS.toMillis(1)); //1s
+            TestHelper.allocGarbage(0);
             numberOfRetries++;
             System.gc();
         }
