+++ realm/src/main/java/io/realm/internal/SharedGroup.java
@@ -24,6 +24,14 @@ import io.realm.exceptions.RealmIOException;
 
 public class SharedGroup implements Closeable {
 
+    public static final boolean IMPLICIT_TRANSACTION = true;
+    public static final boolean EXPLICIT_TRANSACTION = false;
+
+    private static final boolean CREATE_FILE_YES = false;
+    private static final boolean CREATE_FILE_NO = true;
+    private static final boolean ENABLE_REPLICATION = true;
+    private static final boolean DISABLE_REPLICATION = false;
+
     private final String path;
     private long nativePtr;
     private long nativeReplicationPtr;
@@ -50,34 +58,27 @@ public class SharedGroup implements Closeable {
     public SharedGroup(String databaseFile) {
         context = new Context();
         path = databaseFile;
-        nativePtr = nativeCreate(databaseFile, Durability.FULL.value, false, false, null);
+        nativePtr = nativeCreate(databaseFile, Durability.FULL.value, CREATE_FILE_YES, DISABLE_REPLICATION, null);
         checkNativePtrNotZero();
     }
 
-    public SharedGroup(String databaseFile, boolean enableImplicitTransactions, Durability durability, byte[] key) {
+    public SharedGroup(String canonicalPath, boolean enableImplicitTransactions, Durability durability, byte[] key) {
         if (enableImplicitTransactions) {
-            nativeReplicationPtr = nativeCreateReplication(databaseFile, key);
+            nativeReplicationPtr = nativeCreateReplication(canonicalPath, key);
             nativePtr = createNativeWithImplicitTransactions(nativeReplicationPtr, durability.value, key);
             implicitTransactionsEnabled = true;
         } else {
-            nativePtr = nativeCreate(databaseFile, Durability.FULL.value, false, false, key);
+            nativePtr = nativeCreate(canonicalPath, Durability.FULL.value, CREATE_FILE_YES, DISABLE_REPLICATION, key);
         }
         context = new Context();
-        path = databaseFile;
+        path = canonicalPath;
         checkNativePtrNotZero();
     }
 
-    public SharedGroup(String databaseFile, Durability durability, byte[] key) {
-        path = databaseFile;
+    public SharedGroup(String canonicalPath, Durability durability, byte[] key) {
+        path = canonicalPath;
         context = new Context();
-        nativePtr = nativeCreate(databaseFile, durability.value, false, false, key);
-        checkNativePtrNotZero();
-    }
-
-    public SharedGroup(String databaseFile, Durability durability, boolean fileMustExist) {
-        path = databaseFile;
-        context = new Context();
-        nativePtr = nativeCreate(databaseFile, durability.value, fileMustExist, false, null);
+        nativePtr = nativeCreate(canonicalPath, durability.value, false, false, key);
         checkNativePtrNotZero();
     }
 
@@ -217,7 +218,7 @@ public class SharedGroup implements Closeable {
     /**
      * Returns the absolute path to the file backing this SharedGroup.
      *
-     * @return Absolute path to the Realm file.
+     * @return Canonical path to the Realm file.
      */
     public String getPath() {
         return path;
@@ -246,7 +247,7 @@ public class SharedGroup implements Closeable {
     private native void nativeRollback(long nativePtr);
     private native long nativeCreate(String databaseFile,
                                      int durabilityValue,
-                                     boolean no_create,
+                                     boolean dontCreateFile,
                                      boolean enableReplication,
                                      byte[] key);
     private native boolean nativeCompact(long nativePtr);
